intern/requirement.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
aiosqlite==0.19.0
annotated-types==0.7.0
anyio==4.10.0
APScheduler==3.10.4
asgiref==3.8.1
attrs==25.3.0
Authlib==1.6.5
bcrypt==4.3.0
beartype==0.22.8
beautifulsoup4==4.14.3
blinker==1.7.0
boto3==1.39.14
botocore==1.39.14
bs4==0.0.2
cachetools==6.2.2
certifi==2025.11.12
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.7
colorama==0.4.6
colorlog==6.9.0
contourpy==1.3.3
cryptography==45.0.5
cycler==0.12.1
cyclopts==4.3.0
decompyle3==3.9.2
diskcache==5.6.3
dj-rest-auth==7.0.2
Django==5.1.6
django-allauth==65.14.0
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
dnspython==2.8.0
docker==7.1.0
docstring_parser==0.17.0
docutils==0.22.3
drf-spectacular==0.29.0
duo_client==5.5.0
email-validator==2.3.0
et_xmlfile==2.0.0
exceptiongroup==1.3.1
fastapi==0.116.1
fastmcp==2.13.3
Flask==3.0.3
flask-cors==6.0.1
Flask-SQLAlchemy==3.1.1
fonttools==4.59.0
frozenlist==1.7.0
greenlet==3.2.3
gunicorn==22.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
idna==3.10
inflection==0.5.1
iniconfig==2.3.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.5.1
jsonschema==4.25.1
jsonschema-path==0.3.4
jsonschema-specifications==2025.9.1
kiwisolver==1.4.8
lxml==6.0.0
markdown-it-py==4.0.0
MarkupSafe==2.1.3
matplotlib==3.10.3
mcp==1.22.0
mdurl==0.1.2
multidict==6.6.3
mysqlclient==2.2.7
nltk==3.9.1
numpy==2.4.1
openapi-pydantic==0.5.1
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==25.0
pandas==3.0.0
paramiko==3.5.1
passlib==1.7.4
pathable==0.4.4
pathvalidate==3.3.1
pillow==11.3.0
platformdirs==4.5.1
pluggy==1.6.0
propcache==0.3.2
psutil==7.0.0
psycopg2==2.9.10
psycopg2-binary==2.9.11
py-key-value-aio==0.3.0
py-key-value-shared==0.3.0
pycparser==2.22
pycryptodome==3.23.0
pydantic==2.11.7
pydantic-settings==2.12.0
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==2.10.1
PyNaCl==1.5.0
pyotp==2.9.0
pyparsing==3.2.3
pyperclip==1.11.0
PySocks==1.7.1
pytest==9.0.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.2
pywin32==311
PyYAML==6.0.2
redis==6.2.0
referencing==0.36.2
regex==2024.11.6
reportlab==4.4.3
requests==2.31.0
rich==14.2.0
rich-rst==1.3.2
rpds-py==0.30.0
s3transfer==0.13.1
schedule==1.2.2
scikit-learn==1.7.1
scipy==1.16.0
secret==0.8
selenium==4.39.0
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8
spark-parser==1.8.9
SQLAlchemy==2.0.41
sqlparse==0.5.3
sse-starlette==3.0.3
starlette==0.47.3
tabulate==0.9.0
tenacity==9.1.2
threadpoolctl==3.6.0
tk==0.1.0
tqdm==4.67.1
trio==0.32.0
trio-websocket==0.12.2
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.1
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.35.0
watchdog==6.0.0
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.0.1
wsproto==1.2.0
xdis==6.1.5
yarl==1.20.1
